requirements.txt
streamlit
pandas
langchain
langchain-groq
langchain-chroma
langchain-community
sentence-transformers
pillow
pysqlite3-binary